third_party/WebKit/LayoutTests/http/tests/inspector/resource-tree/resource-tree-no-xhrs-expected.txt

CONSOLE WARNING: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
CONSOLE WARNING: Setting 'XMLHttpRequest.withCredentials' for synchronous requests is deprecated.
CONSOLE MESSAGE: line 6: XHR loaded: 1
Tests that XHRs are not added to resourceTreeModel.

Bug 60321
